Refactor agent tests
Split agent_test.go into several smaller files as it was getting quite
large.
